Desktop: Fix Build Warning.
Fix a warning when building with user manual support.

+++ desktop-widgets/usermanual.cpp
@@ -156,7 +156,7 @@ UserManual::UserManual(QWidget *parent) : QDialog(parent)
 	// Size the window to 80% of the main window's size
 	if (parent) {
 		QSize parentSize = parent->size();
-		resize(parentSize.width() * 0.8, parentSize.height() * 0.8);
+		resize(static_cast<int>(parentSize.width() * 0.8), static_cast<int>(parentSize.height() * 0.8));
 	} else {
 		// Fallback size if no parent
 		resize(700, 500);
